requirements.txt
canvasapi
faker
lxml
python-dotenv